pyproject.toml
[project]
name = "storyverse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "azure-cosmos>=4.9.0",
    "azure-storage-blob>=12.25.1",
    "fastapi>=0.115.12",
    "openai>=1.78.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]
